requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
neat-python==0.92
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
urllib3==1.26.9
